Fixed Saas Pragma & misc ui

+++ AL Company Export-Import/src/Codeunit/TOOWriteInsideTryFunctionEnabled.Codeunit.al
@@ -2,18 +2,20 @@ codeunit 51000 "TOO WriteInsideTryEnabled"
 {
     trigger OnRun()
     var
+        EnvironmentInformation: Codeunit "Environment Information";
         Buffer: Record "Name/Value Buffer";
     begin
-#if ONPREM
+        if EnvironmentInformation.IsSaaS() then
+            exit; // Write inside tryfunction is alway enabled on SaaS
+
+        // Test write inside a tryfunction
         Buffer.DeleteAll();
         Buffer.Init();
         Buffer.ID := 1892256;
         if not TryInsert(Buffer) then
             Error('');
         Buffer.Delete();
-#else
-        exit; // Write inside tryfunction is alway enabled on SaaS
-#endif
+        exit;
     end;
 
     [TryFunction]
